requirements.txt
pandas
requests
python-dotenv
tqdm
pyodbc